requirements.txt
numpy
pandas
tensorflow-addons
opencv-python
tqdm
matplotlib
pytest
jupyter
tensorflow-datasets
